pyproject.toml
[tool.poetry]
name = "flask-goit"
version = "0.1.0"
description = ""
authors = ["OleksiiLatypov  <latypov.oleksii.la@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_goit"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.0"
marshmallow = "^3.19.0"
flask-migrate = "^4.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
